update: update ehdr.hpp declare types

+++ include/libelfxx/ehdr.hpp
@@ -4,17 +4,17 @@
 
 struct Ehdr
 {
-  uint16_t e_type;
-  uint16_t e_machine;
-  uint32_t e_version;
-  uint32_t e_entry;
-  uint32_t e_phoff;
-  uint32_t e_shoff;
-  uint32_t e_flags;
-  uint16_t e_ehsize;
-  uint16_t e_phentsize;
-  uint16_t e_phnum;
-  uint16_t e_shentsize;
-  uint16_t e_shnum;
-  uint16_t e_shstrndx;
+  uint e_type;
+  uint e_machine;
+  uint e_version;
+  uint e_entry;
+  uint e_phoff;
+  uint e_shoff;
+  uint e_flags;
+  uint e_ehsize;
+  uint e_phentsize;
+  uint e_phnum;
+  uint e_shentsize;
+  uint e_shnum;
+  uint e_shstrndx;
 };
